Add files via upload
The main page is web0.html. It is a simple login page which is not a proper functioning login page. I've just added some links to other pages. 

After clicking on sign-up, you'll head to the HOMEPAGE. In homepage, I've just simply made a NAVBAR(this you'll find in every page except the login page) and a transition effect with 5-6 images. 

Moving further in navbar you'll reach the ABOUT-US page. Here, I've tried to give a different look to the user and the three buttons are hoverable, their background color changes when you hover over them.

The fourth page is about SERVICES. It gives information about all the services provided by us. I've added a hoverable effect on each service and the content-display I used is flex.

Fifth page is about PROJECTS. Here I've added some random text and some images. There are two parts here - HEADER and FOOTER. The only difference here is the size of images.

And the last page is FEEDBACK. I've created a general feedback form. Every box is given a transparent background with a glowing box-shadow. Also the send button contains hoverable effect.

Every page is designed keeping in mind the device width. This project can smoothly run on any device (phone, tablet or laptop) without cropping or distorting any of the pages designs.